Advanced Code Solution (most performant)
the person who submitted this was given the same problem in an interview (spoiler: he wasn’t hired :(  ) he quickly arrived at the basic solution, and the interviewer told him to make it better. The algorithm would take way too long if he passed the Bible as the string. He wanted it to be instant.

The simpler solutions perform very poorly on long strings because they operate on the whole string multiple times (toLowerCase(), replace(), split(), reverse(), join()) before comparing the whole string twice.

The beauty of this solution is it never needs to read through the whole string, even once, to know that it’s not a palindrome. Why read through the whole string if you can tell that it’s not a palindrome just by looking at two letters?

Uses a while loop instead of a for loop as a best practice - because we are using two variables, one is the index starting from the beginning of the string, and the other starts at the end of the string.